Fixed printing (#9)
* printing no wait time yet w/ bug

Added printing so same process burst prints on same line, getting a bug where its not appending to the first process. seems to work fine for processes other than 1; might be because i removed the loop that checks through the entire list?

Will do the sum once i get these to print out correctly.

* fixed bug

Fixed the bug mentioned in my last push where process 1 was not concatenating properly. This version is also able to process IDELs and print them properly.

Need to do: Summation of wait times and print them on the same line as the process it corresponds to. :D

* Update proc.py

Added a | divider

* fixed printing

Fixed printing finally and the wait times shown is the total of the wait times :)

Tested it out with what sir gave us and it seems correct but his doesn't have the IDLE times printed out.

* Added the AveWT

---------

Co-authored-by: Miguel Escalona - DLSU <71485171+jm55DLSU@users.noreply.github.com>