New Test files (By Mollen Villavicencio) (#12)
* printing no wait time yet w/ bug

Added printing so same process burst prints on same line, getting a bug where its not appending to the first process. seems to work fine for processes other than 1; might be because i removed the loop that checks through the entire list?

Will do the sum once i get these to print out correctly.

* fixed bug

Fixed the bug mentioned in my last push where process 1 was not concatenating properly. This version is also able to process IDELs and print them properly.

Need to do: Summation of wait times and print them on the same line as the process it corresponds to. :D

* Update proc.py

Added a | divider

* fixed printing

Fixed printing finally and the wait times shown is the total of the wait times :)

Tested it out with what sir gave us and it seems correct but his doesn't have the IDLE times printed out.

* Added the AveWT

* added print to text file

Output will print on cmd and output.txt which will show up in the folder :smile_cat:

* Update utils.py

forgot to put the ave wait time haha

* Test Files

Tested the other files and added one new test files for each algo. All goods ang mga calculation.

---------

Co-authored-by: Gabrielle Kaye <75007380+gabriellekaye@users.noreply.github.com>
Co-authored-by: Gabrielle Kaye <gabrielle_naval@dlsu.edu.ph>
Co-authored-by: Mollen26 <82302488+Mollen26@users.noreply.github.com>